implementación del ERMapper y actualización de su DAO

+++ src/main/java/repository/dao/EscapeRoomDAO.java
@@ -1,9 +1,11 @@
 package repository.dao;
 
-import com.mysql.cj.jdbc.exceptions.OperationNotSupportedException;
+import exception.EmptyNameException;
 import exception.PersistenceException;
 import model.EscapeRoom;
 import repository.database.DatabaseConfig;
+import repository.mapper.EscapeRoomMapper;
+import repository.mapper.GeneralMapper;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -13,15 +15,19 @@ import java.util.*;
 
 public class EscapeRoomDAO implements GenericDAO<EscapeRoom, Long> {
 
+    private final GeneralMapper<EscapeRoom> mapper = EscapeRoomMapper.getInstance();
+
     @Override
     public EscapeRoom save(EscapeRoom escapeRoom) {
+        mapper.validateEntity(escapeRoom);
         String sql = "INSERT INTO escape_room (name) VALUES (?)";
 
         try (Connection conn = DatabaseConfig.getConnection();
              PreparedStatement stmt = conn.prepareStatement(sql, PreparedStatement.
                      RETURN_GENERATED_KEYS)) {
 
-            stmt.setString(1, escapeRoom.getName());
+            mapper.toPreparedStatement(escapeRoom, stmt);
+
             int affectedRows = stmt.executeUpdate();
 
             if (affectedRows == 0) {
@@ -29,42 +35,61 @@ public class EscapeRoomDAO implements GenericDAO<EscapeRoom, Long> {
             }
 
             setGeneratedId(escapeRoom, stmt);
-            return escapeRoom;
 
         } catch (SQLException e) {
             throw new PersistenceException("Error al guardar el Escape Room: " + escapeRoom.getName() + ".");
         }
+        return escapeRoom;
     }
-
 
     @Override
     public Optional<EscapeRoom> findById(Long id) {
+        if (id == null || id <= 0) {
+            throw new IllegalArgumentException("El ID debe ser un número positivo.");
+        }
+
         String sql = "SELECT id, name FROM escape_room WHERE id = ?";
+        Optional<EscapeRoom> escapeRoom = Optional.empty();
 
         try (Connection conn = DatabaseConfig.getConnection();
              PreparedStatement stmt = conn.prepareStatement(sql)) {
 
             stmt.setLong(1, id);
-            return executeQueryAndMapToEscapeRoom(stmt);
 
+            try (ResultSet rs = stmt.executeQuery()) {
+                if (rs.next()) {
+                    escapeRoom = Optional.of(mapper.fromResultSet(rs));
+                }
+            }
         } catch (SQLException e) {
             throw new PersistenceException("Error al buscar el Escape Room con ID = " + id + ".");
         }
+        return escapeRoom;
     }
 
     @Override
     public Optional<EscapeRoom> findByName(String name) {
+        if (name == null || name.trim().isEmpty()) {
+            throw new EmptyNameException();
+        }
+
         String sql = "SELECT id, name FROM escape_room WHERE name = ?";
+        Optional<EscapeRoom> escapeRoom = Optional.empty();
 
         try (Connection conn = DatabaseConfig.getConnection();
              PreparedStatement stmt = conn.prepareStatement(sql)) {
 
             stmt.setString(1, name);
-            return executeQueryAndMapToEscapeRoom(stmt);
 
+            try (ResultSet rs = stmt.executeQuery()) {
+                if (rs.next()) {
+                    escapeRoom = Optional.of(mapper.fromResultSet(rs));
+                }
+            }
         } catch (SQLException e) {
             throw new PersistenceException("Error al buscar Escaper Room por nombre: " + name + ".");
         }
+        return escapeRoom;
     }
 
 
@@ -78,30 +103,33 @@ public class EscapeRoomDAO implements GenericDAO<EscapeRoom, Long> {
              ResultSet rs = stmt.executeQuery()) {
 
             while (rs.next()) {
-               EscapeRoom escapeRoom = mapResultSetToEscapeRoom(rs);
-               escapeRooms.add(escapeRoom);
+                escapeRooms.add(mapper.fromResultSet(rs));
             }
-            return escapeRooms;
-
         } catch (SQLException e) {
             throw new PersistenceException("Error al buscar todos los Escape Rooms.");
         }
+        return escapeRooms;
     }
 
     @Override
     public boolean delete(Long id) {
+        if (id == null || id <= 0) {
+            throw new IllegalArgumentException("El ID debe ser un número positivo.");
+        }
+
         String sql = "DELETE FROM escape_room WHERE id = ?";
+        int affectedRows;
 
         try (Connection conn = DatabaseConfig.getConnection();
              PreparedStatement stmt = conn.prepareStatement(sql)) {
 
             stmt.setLong(1, id);
-            int affectedRows = stmt.executeUpdate();
-            return affectedRows > 0;
+            affectedRows = stmt.executeUpdate();
 
         } catch (SQLException e) {
             throw new PersistenceException("Error al borrar el escape room.");
         }
+        return affectedRows > 0;
     }
 
     private void setGeneratedId(EscapeRoom escapeRoom, PreparedStatement stmt) throws SQLException {
@@ -109,25 +137,8 @@ public class EscapeRoomDAO implements GenericDAO<EscapeRoom, Long> {
             if (generatedKeys.next()) {
                 escapeRoom.setId(generatedKeys.getLong(1));
             } else {
-                throw new OperationNotSupportedException("Error al intentar recuperar el ID generado para el Escape Room.");
+                throw new PersistenceException("No se pudo obtener el ID generado para el Escape Room.");
             }
         }
     }
-
-    private Optional<EscapeRoom> executeQueryAndMapToEscapeRoom(PreparedStatement stmt) throws SQLException {
-        try (ResultSet rs = stmt.executeQuery()) {
-            if (rs.next()) {
-                return Optional.of(mapResultSetToEscapeRoom(rs));
-            }
-            return Optional.empty();
-        }
-    }
-
-    private EscapeRoom mapResultSetToEscapeRoom(ResultSet rs) throws SQLException {
-        EscapeRoom escapeRoom = new EscapeRoom(rs.getString("name"));
-        escapeRoom.setId(rs.getLong("id"));
-        return escapeRoom;
-    }
-
-
 }
